dam-automation/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dam-automation"
version = "0.1.0"
description = "Data management automation service for lakehouse resources"
readme = "README.md"
authors = [{name = "Iceberg Automation Team"}]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5",
    "PyYAML>=6.0",
    "requests>=2.31",
    "typer>=0.9",
    "tenacity>=8.2",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.15",
    "azure-mgmt-storage>=21.0",
    "azure-mgmt-msi>=7.0",
    "azure-graphrbac>=0.61",
]

[project.scripts]
dam-automation = "dam_automation.main:app"

[tool.setuptools.packages.find]
where = ["src"]
